更新

+++ GetMsg/loginwindow.py
@@ -14,7 +14,7 @@ import mainwindow
 from newthread import MyThread
 from function.MainWindow import *
 #窗口
-from function.FramelessWindow import FramelessWindow
+from function.FramelessMainWindow import FramelessMainWindow
 import os
 import os.path
 # 获取Windows平台临时文件夹
@@ -23,37 +23,31 @@ filename = os.path.join(path, 'info.txt')
 
 window=tk.Tk()
 window.title('欢迎使用九龙湖邮管队')
-window.geometry('720x400+500+200')
+window.geometry('420x550+500+100')
 window.resizable(0,0)
+window.iconbitmap('title.ico')
 #画布放置图片
 window.resizable(0,0)
-frmlt=tk.Frame(window,width=360,height=450,bg='black')
-frmlt.grid(row=0,column=0)
-canvas1=tk.Canvas(window,height=450,width=360)
-imagefile1=tk.PhotoImage(file='fina.png')
-image=canvas1.create_image(0,-20,anchor='nw',image=imagefile1)
-canvas1.grid(row=0,column=0)
-canvas2=tk.Canvas(window,height=450,width=360)
-imagefile2=tk.PhotoImage(file='3.png')
-image=canvas2.create_image(0,0,anchor='nw',image=imagefile2)
-canvas2.grid(row=0,column=1)
-window.wm_attributes('-topmost',1)
+tk.Label(window,text='Spam',fg='white',anchor='n',bg='FireBrick',font=('Ink Free',50),width=420,height=550).pack()
+imageMail=tk.PhotoImage(file='mail.png')
+imageLock=tk.PhotoImage(file='lock.png')
+tk.Label(window,image=imageMail).place(x=50,y=125)
+tk.Label(window,image=imageLock).place(x=55,y=200)
 #标签 用户名密码
 #用户名输入框
 var_usr_name=tk.StringVar(window,value='')
-entry_usr_name=tk.Entry(window,textvariable=var_usr_name,bg='white')
-entry_usr_name.place(x=360,y=78)
+entry_usr_name=tk.Entry(window,textvariable=var_usr_name,font=('Arial', 30),relief='solid',bg='DarkRed',width=10,bd=1)
+entry_usr_name.place(x=100,y=120)
 #密码输入框
 var_usr_pwd=tk.StringVar(window,value='')
-entry_usr_pwd=tk.Entry(window,textvariable=var_usr_pwd,show='*',bg='white')
-entry_usr_pwd.place(x=510,y=78)
-def pwd_reme():
-   if(var.get()==1):
+entry_usr_pwd=tk.Entry(window,textvariable=var_usr_pwd,show='*',font=('Arial', 30),bg='DarkRed',width=10,bd=1)
+entry_usr_pwd.place(x=100,y=200)
+try:
         with open(filename,'r') as fp:
            n, p = fp.read().strip().split(',')
            var_usr_name.set(n)
            var_usr_pwd.set(p)
-   else:
+except:
       pass     
 
 #登录函数
@@ -73,15 +67,16 @@ def usr_log_in():
     mailusr=MailUser(usr_name,usr_pwd)
     sta = mailusr.login(usr_name,usr_pwd)
     if sta == "not connected":
-            tk.messagebox.showerror(title='警告',message='请检查账号格式或网络连接状态')
+            tk.messagebox.showerror(title='警告',message='请检查账号或网络连接状态')
     elif sta == 'login fail':
             tk.messagebox.showerror(message='账号或密码错误')
 
     elif sta == 'login success':
 
             tk.messagebox.showinfo(title="欢迎",message="登录成功")
-            with open(filename, 'w') as fp:
-              fp.write(','.join((usr_name,usr_pwd)))
+            if usr_command==1:
+             with open(filename, 'w') as fp:
+               fp.write(','.join((usr_name,usr_pwd)))
             window.destroy()
 
             app = mainwindow.QtWidgets.QApplication(sys.argv)
@@ -99,7 +94,7 @@ def usr_log_in():
                     send=mailusr.mailsender(i)
                     clisock.sendfrom(usr_name,send)
                     rs=clisock.get_sender()
-                    #print(rs)
+                    print(rs)
                     if rs=='1':
                         mailusr.setlabel(i, False)
                     elif rs=='2':
@@ -114,14 +109,15 @@ def usr_log_in():
                 else:
                     mailusr.setlabel(i, result)
 
+            ui = FramelessMainWindow(mailusr, clisock)
+
 
 
             signal = SiganlObj()
-            slot = TypeSlot()
+            slot = TypeSlot(ui,mailusr)
             signal.sendMsg.connect(slot.get)
 
 
-            ui = FramelessWindow(mailusr, clisock)
 
             ui.mainwindow.setUserName(usr_name)
             ui.show()
@@ -134,15 +130,21 @@ def usr_log_in():
     
 def usr_sign_quit():
     window.destroy()
+    
+def usr_command():
+    if(var.get()==1):
+        return 1
+    else:
+        return 0    
 
 #登录 按钮
-bt_login=tk.Button(window,text='登录',command=usr_log_in)
-bt_login.place(x=140,y=230)
-bt_logquit=tk.Button(window,text='退出',command=usr_sign_quit)
-bt_logquit.place(x=280,y=230)
+bt_login=tk.Button(window,text='Login',width='16',height='1',font=('Ink Free', 17),anchor='center',cursor='hand2',command=usr_log_in,bg='LightCyan')
+bt_login.place(x=100,y=310)
+bt_logquit=tk.Button(window,text='Exit',width='16',height='1',font=('Ink Free', 17),anchor='center',cursor='hand2',command=usr_sign_quit,bg='LightCyan')
+bt_logquit.place(x=100,y=390)
 var=tk.IntVar()
-bt_reme =tk.Checkbutton(window,text='记录上次登录信息',variable=var,onvalue=1,offvalue=0,command=pwd_reme)
-bt_reme.place(x=360,y=105)
+bt_reme =tk.Checkbutton(window,text='记住密码',bg='FireBrick',variable=var,onvalue=1,offvalue=0,command=usr_command)
+bt_reme.place(x=100,y=250)
 #主循环
 window.mainloop()
